+++ src/django_subatomic/test.py
@@ -3,15 +3,31 @@ from __future__ import annotations
 import contextlib
 from typing import TYPE_CHECKING
 
+import attrs
+from django.conf import settings
 from django.db import transaction
 
 
 if TYPE_CHECKING:
-    from collections.abc import Generator
+    from collections.abc import Callable, Generator
 
 __all__ = [
     "part_of_a_transaction",
 ]
+
+
+@attrs.frozen
+class _UnhandledCallbacks(Exception):
+    """
+    Raised in tests when unhandled callbacks are found before calling `part_of_a_transaction`.
+
+    This happens when after-commit callbacks are registered
+    but not run before trying to open a database transaction.
+
+    The best solution is to ensure the after-commit callbacks are handled first.
+    """
+
+    callbacks: tuple[Callable[[], object], ...]
 
 
 @contextlib.contextmanager
@@ -33,6 +49,12 @@ def part_of_a_transaction(using: str | None = None) -> Generator[None]:
     use [`transaction`][django_subatomic.db.transaction] instead.
     """
     connection = transaction.get_connection(using)
+    if getattr(
+        settings, "SUBATOMIC_CATCH_UNHANDLED_AFTER_COMMIT_CALLBACKS_IN_TESTS", True
+    ):
+        callbacks = connection.run_on_commit
+        if callbacks:
+            raise _UnhandledCallbacks(tuple(callback for _, callback, _ in callbacks))
 
     with transaction.atomic(using=using, durable=True):
         yield